Fix bug 'delete task' bug
Turns out, circular references WERE a bad idea, because of course they immediately broke the JSON parser.
The "parent" object (project) had a reference to its "child" (task), & the child had a reference to the parent.
The only thing I was using the parent reference for was to delete, which just removed itself from its parent's children.
Happily, all of that has been removed.
No, I'm still not willing to add a C to my MVC, sorry :I